api/requirements.txt
azure-functions==1.21.3
azure-communication-email==1.0.0
azure-data-tables==12.5.0
azure-mgmt-appcontainers==3.1.0
azure-identity==1.19.0
httpx==0.27.2
pyjwt==2.9.0
pydantic==2.9.2
